add support for cells with intervals
modify `Cell` class to have a method to merge new content
with the existing one, to be overidden by classes that inherit it.

add a `Interval` class to represent number intervals, and a
`IntervalCell` that inherits `Cell` and allows it to hold intervals as
content and merge them with other intervals when they are added as
content.

add new primitives (squarer and sqrter) that work with intervals