add operator module to hold generic operators
these operators will be used throughout all the `propagator` module
instead of python's traditional arithmetic operators.

this decision was made to ease the implementation of operations between
different types of variables (e.g. `Supported` and `Interval`, or
`Interval` and flat numbers), in a way that allows each module that
implements a new type (e.g. `propagator.interval`) to also define how
to handle operations between the new type and other types.

each one of the implemented operators defaults to a python traditional
operator (taken from the `operator` python module), but can be
extended to handle unanticipated types in the future.