test for 'nothing' instead of 'none' when merging
when merging two cell contents, there was an operation assigned for the
case where one argument is `None` and the other is not.

this should really be a test for 'nothingness', for Supported and TMS
values can 'wrap' `None` values (e.g. `Supported(None, [])`.

instead of using `is_none` and `is_not_none` functions (which are wrappers
around `is None` and `is not None`), it uses `is_nothing` and
`is_not_nothing`, which are generic operators that default to the former
functions but can be overidden.