search for generic operation from last added to first.
append new operations on the left of the list, instead of the right.

since it searches for the operation for a certain list of arguments from
left to right, newly added operations will be tested first.

this gives us the ability to override general operations (which are
added first) with more specific ones